move the bq config again
