+++ src/main/resources/application.properties
@@ -4,7 +4,7 @@ spring.datasource.tomcat.test-on-borrow=true
 spring.datasource.tomcat.validation-query=SELECT 1
 spring.jackson.serialization.WRITE_DATES_AS_TIMESTAMPS=false
 spring.liquibase.enabled=false
-spring.profiles.active=dev
+spring.profiles.active=google
 spring.resources.cache.cachecontrol.max-age=0
 spring.resources.cache.cachecontrol.must-revalidate=true
 spring.resources.static-locations=classpath:/api/
